Update SpecifiedInputBashScript.sh
Parallelising the watcher for each simulation run. Allows for all simulations to be ran in parallel. In addition any single simulation that fails due to reaching the timeout of 1800s (30 minutes) will be killed due to hanging. 

Updating the "sleep 1800" value (in seconds) will change the tolerance of how long we enjoy a set of simulations to run. e.g for simulations running at an average of 35 minutes to complete you would need to increase the sleep to accomodate this.