Alternate Bash Script for hanging simulations
Adding a alternate bash script pair for possible hanging simulations in advance of a permanent fix. These files are commented.

If a test being ran is found to not be updating its test file through checkusage.sh  in a given amount of time then a kill signal will be sent to its parent process test and an error message will be output.

Do not run checkusage.sh by itself as it has no default parameters.